rename to setWeight
