+++ client/src/functions/setWeightValue.js
